Separate pipeline and evaluation
Cannot be fully separated easily, because there are some pipeline routines
which already create plots. Also, class Controller has a state, so
independently executing some of its methods is tricky. As we don't want to
do a major refactoring at the moment, we separate pipeline and evaluation
as good as we can without much effort.

Addresses issue #33
Resolves issue #42